% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(x)
}
\arguments{
\item{x}{is the input.}
}
\value{
is the vector of integers counting up to x from 0, as well as the strings that replace integers divisible by 3 or 5.
}
\description{
FizzBuzz takes one input and returns a vector of numbers counting up to the input, from 0, however, if the number can be divided by 3, 5, or 3 and 5, it displays the strings Fizz, Buzz, or FizzBuzz.
}
