% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of different weather data in Flagstaff, AZ.}
\format{
a data frame with 25,449 observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{PRCP}{The amount of precipitation of an observation in inches}
  \item{SNOW}{The amount of snow of an observation in inches.}
  \item{TMAX}{The maximum temperature of an observation in fahrenheit.}
  \item{TMIN}{The minimum temperature of an observation in fahrenheit.}
  \item{SNWD}{The depth of snow of an observation in inches.}
}
}
\source{
\url{www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of different weather data in Flagstaff, AZ.
}
\keyword{datasets}
